+++ admin-ui/src/App.tsx
@@ -30,8 +30,8 @@ import SpaceEditorView from "./component/Spaces/space-editor/SpaceEditorView";
 import ResourceEditorView from "./component/Resources/resource-editor/ResourceEditorView";
 import ReservationUnitEditor from "./component/ReservationUnits/ReservationUnitEditor/ReservationUnitEditor";
 import ResourcesList from "./component/Resources/ResourcesList";
-import ReservationUnitsList from "./component/ReservationUnits/ReservationUnitsList";
-import ReservationUnitsSearch from "./component/ReservationUnits/ReservationUnitsSearch";
+import ReservationUnitsList from "./component/reservation-units/ReservationUnitsList";
+import ReservationUnitsSearch from "./component/reservation-units/ReservationUnitsSearch";
 import { withGlobalContext } from "./context/GlobalContexts";
 
 import RequestedReservations from "./component/reservations/requested/RequestedReservations";
